move dag's methods into dag.go
